requirements.txt
openai==2.14.0
httpx==0.28.1
httpcore==1.0.9
python-dotenv==1.0.1
